unoplat-code-confluence-query-engine/pyproject.toml
[project]
name = "unoplat-code-confluence-query-engine"
version = "0.42.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiofiles>=24.0.0",
    "aiopath>=0.7.7",
    "asyncpg>=0.30.0",
    "fastapi[standard]>=0.116.1",
    #"fastmcp>=2.14.0",
    "ghapi>=1.0.10",
    "jinja2>=3.1.6",
    "logfire[fastapi]>=4.4.0",
    "loguru>=0.7.3",
    "markdown-it-py>=4.0.0",
    "pydantic-ai>=1.86.1",
    "pydantic-ai-backend[docker]>=0.2.4",
    "pydantic-ai-skills>=0.1.0",
    "pydantic-ai-slim[bedrock,duckduckgo,web-fetch,retries]>=1.83.0",
    "qs-codec>=1.1.1",
    "sqlalchemy[asyncio]>=2.0.0",
    "sqlmodel>=0.0.24",
    "sse-starlette>=3.0.0",
    "unoplat-code-confluence-commons>=0.25.0",
]

[dependency-groups]
test = [
    "pytest>=8.0.0",
    "pytest-cov>=4.0.0",
    "pytest-asyncio>=0.23.0",
    "httpx>=0.27.0",
    "docker>=6.0.0",
    "testcontainers>=3.7.0",
    "requests>=2.28.0",
    "python-dotenv>=1.0.0",
    "psycopg2-binary>=2.9.10"
]
dev = [
    "basedpyright>=1.32.1",
    "ruff>=0.1.0",
    "pytest>=8.0.0",
    "pytest-cov>=4.0.0",
    "pytest-asyncio>=0.23.0",
    "types-aiofiles>=24.1.0.20250801",
]

[tool.pytest.ini_options]
pythonpath = ["src"]
# Register custom markers to silence unknown-mark warnings
markers = [
    "integration: integration tests connecting to external services",
]

[tool.uv.sources]
#NOTE: use it for local development to instantly see changes in the schema and its impact
#unoplat-code-confluence-commons = { path = "../unoplat-code-confluence-commons" }
unoplat-code-confluence-commons = { git = "https://github.com/unoplat/unoplat-code-confluence.git", subdirectory = "unoplat-code-confluence-commons", rev = "unoplat-code-confluence-commons-v0.47.0" }

[tool.basedpyright]
typeCheckingMode = "strict"
include = ["src"]
exclude = ["**/__pycache__", ".venv", "dist", "build"]
ignore = ["**/__pycache__"]
useLibraryCodeForTypes = true
reportMissingImports = "error"
reportMissingTypeStubs = false
reportUnusedImport = "warning"
reportUnusedVariable = "warning"
